Typescript defs: fix to be generic JSX
Typescript has an issue with defining the return type of a component function as React.Node, so JSX.Element is used instead which is a better generic definition that should work with non-React JSX type definitions.

Similarly the return type of the callback may be scalar children or complex structures, so updated accordingly while avoiding direct use of the React types.

Tested with CRA --typescript

Fixes #19